Time: 3 ms (45.41%), Space: 8.6 MB (40.88%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         // Negative numbers are never palindrome
         if (x < 0) return false;
 
-        int reversed = 0;
+        long reversed = 0;
         int original = x;
 
         while (x != 0) {
